pyproject.toml
[project]
name = "mbml-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "matplotlib>=3.10.9",
    "numpy>=2.4.4",
    "pandas>=3.0.2",
    "pyro-ppl>=1.9.1",
    "requests>=2.33.1",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "torch>=2.11.0",
    "trueskill>=0.4.5",
]
